reorganizing run files
